requirements.txt
# TutorAI Requirements - 版本兼容修正
# 解决transformers和tokenizers的版本冲突

# ==================== 核心框架 ====================
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
requests==2.31.0
python-multipart==0.0.6

# ==================== 数据处理 ====================
numpy==1.24.3
pandas==2.1.4
pydantic==2.5.0
sqlalchemy==2.0.23

# ==================== 配置管理 ====================
PyYAML==6.0.1
python-dotenv==1.0.0

# ==================== 文档处理 ====================
PyMuPDF==1.23.8
python-docx==1.1.0
markdown==3.5.1

# ==================== 机器学习 - 兼容版本组合 ====================
# 使用兼容的版本组合，避免依赖冲突
transformers==4.35.0
huggingface-hub==0.17.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# ==================== 安装说明 ====================
#
# 推荐安装步骤:
#
# 1. 清理环境:
#    pip cache purge
#    pip freeze | xargs pip uninstall -y
#
# 2. 升级基础工具:
#    pip install --upgrade pip setuptools wheel
#
# 3. 安装PyTorch (必须先安装):
#    pip install torch --index-url https://download.pytorch.org/whl/cpu
#
# 4. 安装其他依赖:
#    pip install -r requirements.txt
#
# 5. 验证安装:
#    python -c "import sentence_transformers, fitz; print('Success!')"
#
# 注意: 已移除tokenizers的版本限制，让transformers自动选择兼容版本